Pybind11

+++ src/NDArrayUtils.h
@@ -219,7 +219,7 @@ std::vector<std::vector<int>> listify(const size_t pop, const NDArray<T>& t, int
       const std::vector<int64_t>& ref = index;
       for (size_t j = 0; j < t.dim(); ++j)
       {
-        list[j][pindex] = offset + ref[/*t.dim() - 1 - */j];
+        list[j][pindex] = offset + static_cast<int>(ref[/*t.dim() - 1 - */j]);
       }
       ++pindex;
     }
